🌭🪡 ↝ [SSM-66 SSG-85]: Removed some old knip files and adjusted durability standard for structures
